review feedback
More avoidance of "tricky" FreeType-ism
